refactor frontend file structure
