+++ frontend/src/components/elements/CurrencySelector.vue
