+++ frontend/src/components/elements/InfoPoint.vue
