+++ frontend/src/components/elements/PlaceElement.vue
