+++ frontend/src/components/elements/ProgressCircle.vue
