+++ frontend/src/components/travel/TravelPage.vue
@@ -396,13 +396,13 @@
 <script lang="ts">
 import { defineComponent, PropType } from 'vue'
 import { Modal } from 'bootstrap'
-import StatePipeline from '../Elements/StatePipeline.vue'
-import ProgressCircle from '../Elements/ProgressCircle.vue'
-import StageForm from '../Forms/StageForm.vue'
-import ExpenseForm from '../Forms/ExpenseForm.vue'
-import InfoPoint from '../Elements/InfoPoint.vue'
-import PlaceElement from '../Elements/PlaceElement.vue'
-import TravelApplyForm from '../Forms/TravelApplyForm.vue'
+import StatePipeline from './elements/StatePipeline.vue'
+import ProgressCircle from '../elements/ProgressCircle.vue'
+import StageForm from './forms/StageForm.vue'
+import ExpenseForm from './forms/ExpenseForm.vue'
+import InfoPoint from '../elements/InfoPoint.vue'
+import PlaceElement from '../elements/PlaceElement.vue'
+import TravelApplyForm from './forms/TravelApplyForm.vue'
 import { getMoneyString, datetoDateString, getLumpSumsSum, getExpensesSum, getTravelTotal } from '../../../../common/scripts.js'
 import { log } from '../../../../common/logger.js'
 import { DocumentFile, TravelExpense, Record, RecordType, Stage, Travel, TravelDay, TravelSimple, meals } from '../../../../common/types.js'
