+++ frontend/src/components/travel/forms/TravelApproveForm.vue
@@ -23,8 +23,8 @@
 
 <script lang="ts">
 import { defineComponent, PropType } from 'vue'
-import TravelApply from '../Elements/TravelApplication.vue'
-import { TravelSimple } from '../../../../common/types.js'
+import TravelApply from '../elements/TravelApplication.vue'
+import { TravelSimple } from '../../../../../common/types.js'
 export default defineComponent({
   name: 'InfoPoint',
   data() {
